move snackbar provider to design-system and provide it to dependents
