--- assets/src/dashboard/app/snackbar/useSnackbarContext.js
@@ -1,33 +1,0 @@
-/*
- * Copyright 2020 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * Internal dependencies
- */
-import { useContextSelector, identity } from '../../../design-system';
-import { SnackbarContext } from './snackbarProvider';
-
-const useSnackbarContext = (selector = identity) => {
-  const context = useContextSelector(SnackbarContext, selector);
-  if (!context) {
-    throw new Error(
-      'useSnackbarContext() must be used within a <SnackbarProvider />'
-    );
-  }
-  return context;
-};
-
-export default useSnackbarContext;
