+++ assets/src/dashboard/app/views/editorSettings/index.js
@@ -37,7 +37,6 @@ import { useConfig } from '../../config';
 import { PageHeading } from '../shared';
 import useTelemetryOptIn from '../shared/useTelemetryOptIn';
 import useMediaOptimization from '../shared/useMediaOptimization';
-import { DashboardSnackbar } from '..';
 import GoogleAnalyticsSettings from './googleAnalytics';
 import { Main, Wrapper } from './components';
 import AdManagement from './adManagement';
@@ -394,9 +393,6 @@ function EditorSettings() {
             )}
           </Main>
         </Layout.Scrollable>
-        <Layout.Fixed>
-          <DashboardSnackbar />
-        </Layout.Fixed>
       </Wrapper>
 
       <Dialog
