+++ assets/src/dashboard/app/views/index.js
@@ -20,4 +20,3 @@ export { default as PreviewStoryView } from './previewStory';
 export { default as TemplateDetailsView } from './templateDetails';
 export { default as SavedTemplatesView } from './savedTemplates';
 export { default as StoryAnimTool } from './storyAnimTool';
-export { default as DashboardSnackbar } from './dashboardSnackbar';
