+++ assets/src/dashboard/app/views/myStories/content/storiesView/index.js
@@ -30,6 +30,7 @@ import {
   Button,
   BUTTON_TYPES,
   BUTTON_SIZES,
+  useSnackbar,
 } from '../../../../../../design-system';
 
 import { StoriesPropType, StoryActionsPropType } from '../../../../../types';
@@ -44,7 +45,6 @@ import {
   STORY_CONTEXT_MENU_ACTIONS,
   STORY_CONTEXT_MENU_ITEMS,
 } from '../../../../../constants';
-import { useSnackbarContext } from '../../../../snackbar';
 import { StoryGridView, StoryListView } from '../../../shared';
 
 const ACTIVE_DIALOG_DELETE_STORY = 'DELETE_STORY';
@@ -68,9 +68,7 @@ function StoriesView({
   const [focusedStory, setFocusedStory] = useState({});
   const [returnStoryFocusId, setReturnStoryFocusId] = useState(null);
 
-  const {
-    actions: { addSnackbarMessage },
-  } = useSnackbarContext();
+  const { showSnackbar } = useSnackbar();
 
   const isActiveDeleteStoryDialog =
     activeDialog === ACTIVE_DIALOG_DELETE_STORY && activeStory;
@@ -160,7 +158,7 @@ function StoriesView({
       setContextMenuId(-1);
       global.navigator.clipboard.writeText(story.link);
 
-      addSnackbarMessage({
+      showSnackbar({
         message:
           story.title.length > 0
             ? sprintf(
@@ -172,10 +170,10 @@ function StoriesView({
                 '(no title) has been copied to your clipboard.',
                 'web-stories'
               ),
-        id: Date.now(),
+        dismissable: true,
       });
     },
-    [addSnackbarMessage]
+    [showSnackbar]
   );
 
   const enabledMenuItems = useMemo(() => {
