+++ assets/src/dashboard/app/views/myStories/content/test/content.js
@@ -21,7 +21,7 @@ import { renderWithProviders } from '../../../../../testUtils';
 
 import { VIEW_STYLE, STORY_STATUSES } from '../../../../../constants';
 import LayoutProvider from '../../../../../components/layout/provider';
-import { SnackbarProvider } from '../../../../snackbar';
+import { SnackbarProvider } from '../../../../../../design-system';
 import Content from '../';
 
 const fakeStories = [
