+++ assets/src/dashboard/app/views/myStories/content/test/storiesView.js
@@ -18,7 +18,7 @@
  * Internal dependencies
  */
 import { renderWithProviders } from '../../../../../testUtils';
-import { SnackbarProvider } from '../../../../snackbar';
+import { SnackbarProvider } from '../../../../../../design-system';
 import {
   STORY_SORT_OPTIONS,
   SORT_DIRECTION,
