+++ assets/src/dashboard/app/views/savedTemplates/content/stories/index.js
@@ -26,7 +26,7 @@ import { FlagsProvider } from 'flagged';
  */
 import { Layout } from '../../../../../components';
 import { VIEW_STYLE } from '../../../../../constants';
-import { SnackbarProvider } from '../../../../snackbar';
+import { SnackbarProvider } from '../../../../../../design-system';
 import {
   formattedStoriesArray,
   formattedTemplatesArray,
