+++ assets/src/design-system/components/index.js
@@ -33,7 +33,7 @@ export * from './radio';
 export * from './search';
 export * from './slider';
 export * from './swatch';
-export { Snackbar } from './snackbar';
+export { Snackbar, SnackbarProvider, useSnackbar } from './snackbar';
 export * from './tooltip';
 export * from './keyboard';
 export * from './keyboard/gridview';
