+++ assets/src/edit-story/app/media/media3p/useFetchMediaEffect.js
@@ -22,7 +22,7 @@ import { useEffect, useRef } from 'react';
 /**
  * Internal dependencies
  */
-import { useSnackbar } from '../../snackbar';
+import { useSnackbar } from '../../../../design-system';
 import { useMedia3pApi } from './api';
 import { PROVIDERS } from './providerConfiguration';
