+++ assets/src/edit-story/app/media/test/useUploadMedia.js
@@ -61,7 +61,7 @@ jest.mock('../utils/useMediaUploadQueue', () => ({
 
 const mockShowSnackbar = jest.fn();
 
-jest.mock('../../snackbar', () => ({
+jest.mock('../../../../design-system/components/snackbar/useSnackbar', () => ({
   useSnackbar: jest.fn(() => ({
     showSnackbar: mockShowSnackbar,
   })),
