+++ assets/src/edit-story/app/story/actions/useSaveStory.js
@@ -29,7 +29,7 @@ import objectPick from '../../../utils/objectPick';
 import { useAPI } from '../../api';
 import { useConfig } from '../../config';
 import useRefreshPostEditURL from '../../../utils/useRefreshPostEditURL';
-import { useSnackbar } from '../../snackbar';
+import { useSnackbar } from '../../../../design-system';
 import getStoryPropsToSave from '../utils/getStoryPropsToSave';
 import { useHistory } from '../../history';
