+++ assets/src/edit-story/components/devTools/devTools.js
@@ -24,9 +24,8 @@ import styled from 'styled-components';
  * Internal dependencies
  */
 import { DATA_VERSION } from '@web-stories-wp/migration';
-import { useGlobalKeyDownEffect } from '../../../design-system';
+import { useGlobalKeyDownEffect, useSnackbar } from '../../../design-system';
 import { useStory } from '../../app/story';
-import { useSnackbar } from '../../app/snackbar';
 import { dummyImage, dummyVideo } from './dummyData';
 
 const Container = styled.div`
