+++ assets/src/edit-story/components/library/panes/media/common/stories/mediaElement.js
@@ -27,7 +27,7 @@ import styled from 'styled-components';
 import MediaElement from '../mediaElement';
 import ApiContext from '../../../../../../app/api/context';
 import MediaContext from '../../../../../../app/media/context';
-import SnackbarContext from '../../../../../../app/snackbar/context';
+import SnackbarContext from '../../../../../../../design-system/components/snackbar/context';
 import testImage from '../../local/stories/test-image.jpg';
 import testPoster from './test-poster.png';
 import testVideo from './test-video.mp4';
