+++ assets/src/edit-story/components/library/panes/media/local/mediaEditDialog.js
@@ -30,7 +30,7 @@ import { useAPI } from '../../../../../app/api';
 import Dialog from '../../../../dialog';
 import { Plain } from '../../../../button';
 import { useLocalMedia } from '../../../../../app/media';
-import { useSnackbar } from '../../../../../app/snackbar';
+import { useSnackbar } from '../../../../../../design-system';
 import StoryPropTypes from '../../../../../types';
 import { getSmallestUrlForWidth } from '../../../../../elements/media/util';
