+++ assets/src/edit-story/editorApp.js
@@ -24,6 +24,7 @@ import PropTypes from 'prop-types';
 /**
  * Internal dependencies
  */
+import { SnackbarProvider } from '../design-system';
 import theme, { GlobalStyle } from './theme';
 import ErrorBoundary from './components/errorBoundary';
 import { ConfigProvider } from './app/config';
@@ -31,7 +32,6 @@ import { APIProvider } from './app/api';
 import { FileProvider } from './app/file';
 import { Media3pApiProvider } from './app/media/media3p/api';
 import { HistoryProvider } from './app/history';
-import { SnackbarProvider } from './app/snackbar';
 import { StoryProvider } from './app/story';
 import { FontProvider } from './app/font';
 import { MediaProvider } from './app/media';
